refactor: Card/IDeckServiceの実体をfeatures/deck/typesに移動（TASK-0068）
domain/entities/Card.tsとdomain/interfaces/deck-service.interface.tsの実体を
features/deck/types/に移動。旧パスは後方互換性のため再エクスポートとして維持。
