+++ atelier-guild-rank/src/domain/interfaces/deck-service.interface.ts
@@ -1,163 +1,10 @@
 /**
- * deck-service.interface.ts - DeckServiceインターフェース
- *
- * TASK-0009: カードエンティティ・DeckService実装
+ * deck-service.interface.ts - DeckServiceインターフェース（後方互換性用再エクスポート）
  *
  * @description
- * デッキサービスの責務を定義するインターフェース。
- * 山札・手札・捨て札の状態管理とカード操作を提供する。
+ * 実体は features/deck/types/deck-service.ts に移動済み。
+ * 後方互換性のため再エクスポートを提供する。
  *
- * @信頼性レベル 🔵
- * - note.mdに基づいた定義
- * - 設計文書に準拠
+ * 新規コードでは @features/deck/types を使用すること。
  */
-
-import type { Card } from '@domain/entities/Card';
-import type { CardId } from '@shared/types';
-
-/**
- * 【機能概要】: DeckServiceインターフェース
- * 【実装方針】: デッキ操作、状態取得、デッキ構築、初期化メソッドを定義
- * 【テスト対応】: T-0009-01 〜 T-DECK-05 を通すための定義
- * 🔵 信頼性レベル: note.md・設計文書に明記
- */
-export interface IDeckService {
-  // =============================================================================
-  // 初期化メソッド
-  // =============================================================================
-
-  /**
-   * 【機能概要】: 初期デッキを構築してシャッフル
-   * 【実装方針】: 指定されたカードIDからCardインスタンスを生成し、デッキに追加してシャッフル
-   * 【テスト対応】: T-0009-01 initialize()で初期デッキ構築
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   *
-   * @param cardIds - 初期デッキのカードID配列
-   */
-  initialize(cardIds: CardId[]): void;
-
-  /**
-   * 【機能概要】: 山札・手札・捨て札をクリア
-   * 【実装方針】: 全ての内部配列を空にする
-   * 【テスト対応】: T-DECK-01 reset()で状態リセット
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   */
-  reset(): void;
-
-  // =============================================================================
-  // デッキ操作メソッド
-  // =============================================================================
-
-  /**
-   * 【機能概要】: 山札をFisher-Yatesアルゴリズムでシャッフル
-   * 【実装方針】: 要素を保持したまま順序のみをランダム化
-   * 【テスト対応】: T-0009-02 shuffle()でランダムにシャッフル
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   */
-  shuffle(): void;
-
-  /**
-   * 【機能概要】: 山札から指定枚数をドロー
-   * 【実装方針】: デッキ枯渇時は捨て札を自動的にシャッフルして山札に戻す
-   * 【テスト対応】: T-0009-03 draw()で手札にカードが追加される、T-0009-04 デッキ枯渇時のドロー
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   *
-   * @param count - ドローする枚数
-   * @returns ドローしたカードの配列
-   */
-  draw(count: number): Card[];
-
-  /**
-   * 【機能概要】: 手札からカードを使用し、捨て札に移動
-   * 【実装方針】: 手札から指定カードを削除し、捨て札に追加
-   * 【テスト対応】: T-0009-05 playCard()で手札から捨て札に移動
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   *
-   * @param card - プレイするカード
-   */
-  playCard(card: Card): void;
-
-  /**
-   * 【機能概要】: 手札を全て捨て札に移動
-   * 【実装方針】: 手札配列の全要素を捨て札に移動し、手札を空にする
-   * 【テスト対応】: T-DECK-02 discardHand()で手札を全て捨て札に移動
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   */
-  discardHand(): void;
-
-  /**
-   * 【機能概要】: 手札が5枚になるまでドロー
-   * 【実装方針】: HAND_SIZE（5枚）まで不足分をドロー
-   * 【テスト対応】: T-0009-06 refillHand()で手札を5枚まで補充
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   */
-  refillHand(): void;
-
-  // =============================================================================
-  // 状態取得メソッド
-  // =============================================================================
-
-  /**
-   * 【機能概要】: 山札の内容を取得（読み取り専用）
-   * 【実装方針】: 内部配列のコピーをreadonly配列として返す
-   * 【テスト対応】: T-DECK-06 getDeck()で山札を取得
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   *
-   * @returns 山札の配列（読み取り専用）
-   */
-  getDeck(): readonly Card[];
-
-  /**
-   * 【機能概要】: 手札の内容を取得（読み取り専用）
-   * 【実装方針】: 内部配列のコピーをreadonly配列として返す
-   * 【テスト対応】: T-DECK-07 getHand()で手札を取得
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   *
-   * @returns 手札の配列（読み取り専用）
-   */
-  getHand(): readonly Card[];
-
-  /**
-   * 【機能概要】: 捨て札の内容を取得（読み取り専用）
-   * 【実装方針】: 内部配列のコピーをreadonly配列として返す
-   * 【テスト対応】: T-DECK-08 getDiscard()で捨て札を取得
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   *
-   * @returns 捨て札の配列（読み取り専用）
-   */
-  getDiscard(): readonly Card[];
-
-  /**
-   * 【機能概要】: 手札の枚数を取得
-   * 【実装方針】: 手札配列の長さを返す
-   * 【テスト対応】: T-DECK-09 getHandSize()で手札枚数を取得
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   *
-   * @returns 手札の枚数
-   */
-  getHandSize(): number;
-
-  // =============================================================================
-  // デッキ構築メソッド
-  // =============================================================================
-
-  /**
-   * 【機能概要】: デッキにカードを追加
-   * 【実装方針】: マスターデータからCardインスタンスを生成し、デッキに追加
-   * 【テスト対応】: T-DECK-04 addCard()でカードをデッキに追加
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   *
-   * @param cardId - 追加するカードID
-   */
-  addCard(cardId: CardId): void;
-
-  /**
-   * 【機能概要】: デッキからカードを削除
-   * 【実装方針】: 指定されたカードIDに一致する最初のカードを削除
-   * 【テスト対応】: T-DECK-05 removeCard()でカードをデッキから削除
-   * 🔵 信頼性レベル: note.md・設計文書に明記
-   *
-   * @param cardId - 削除するカードID
-   */
-  removeCard(cardId: CardId): void;
-}
+export type { IDeckService } from '@features/deck/types/deck-service';
